@@ -1,9 +1,9 @@
 name: zimple
 description: Zimple projektverktyg
-version: 1.2.1+17
+version: 1.2.5+20
 
 environment:
-  sdk: ">=2.19.0 <3.0.0"
+  sdk: ">=3.0.0 <3.1.0"
 
 dependencies:
   clipboard: ^0.1.3
@@ -29,6 +29,7 @@ dependencies:
     sdk: flutter
   flutter_svg: ^2.0.5
   fluttericon: ^2.0.0
+  flutter_pdfview: ^1.3.1
   font_awesome_flutter: ^10.4.0
   focus_detector: ^2.0.0
   image_picker: ^0.8.4+4
